+++ CHANGELOG.md
@@ -9,9 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Support for displaying speed at which the CPU is running in web debug
-* Headless execution mode in Boytacean SDL
-* Many more parameters added for Boytacean SDL
+*
 
 ### Changed
 
@@ -20,6 +18,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 *
+
+## [0.9.4] - 2023-06-04
+
+### Added
+
+* Support for displaying speed at which the CPU is running in Web mode, for debug purposes
+* Headless execution mode in Boytacean SDL
+* Many more parameters added for Boytacean SDL
 
 ## [0.9.3] - 2023-05-18